Rename json related packages
